[6] refactored the sidedrawer to be reusable
