+++ BodyPositiveZumba/Home/InitialHome/HomeUI/HomeView.swift
@@ -8,30 +8,18 @@
 import SwiftUI
 
 struct HomeView: View {
-    @State private var viewModel = HomeViewModel()
+
+    @State private var homeViewModel = HomeViewModel()
+    @StateObject private var sideDrawerViewModel = SideDrawerViewModel()
 
     var body: some View {
         ZStack {
             VStack {
-                HStack {
-                    Button(action: {
-                        viewModel.toggleMenu()
-                    }) {
-                        Image(systemName: Constants.Home.buttonImage)
-                            .resizable()
-                            .frame(width: 24, height: 18)
-                            .foregroundColor(.gray)
-                            .padding()
-                    }
+                TopBarView(buttonAction: {
+                    sideDrawerViewModel.toggleMenu()
+                }, logoName: Constants.Common.logoName)
 
-                    Spacer()
-                    Image(Constants.Common.logoName)
-                        .resizable()
-                        .scaledToFit()
-                        .frame(height: 50)
-                }
-                .frame(maxWidth: .infinity, alignment: .center)
-
+                // Main content
                 ScrollView {
                     VStack {
                         Text(Constants.Home.grettingMessage)
@@ -50,79 +38,7 @@ struct HomeView: View {
                     }
                 }
             }
-
-            Color.black.opacity(
-                viewModel.viewState.isMenuOpen ? Constants.SideDrawer.backgroundOpacity : 0
-            )
-                .ignoresSafeArea()
-                .onTapGesture {
-                    viewModel.closeMenu()
-                }
-                .animation(
-                    .easeInOut(duration: Constants.SideDrawer.animationDuration),
-                    value: viewModel.viewState.isMenuOpen)
-
-            HStack {
-                VStack(
-                    alignment: .leading, spacing: Constants.SideDrawer.buttonSpacing) {
-                    Spacer()
-                    DrawerButton(
-                        title: Constants.SideDrawer.joinNowText,
-                        icon: Constants.SideDrawer.joinNowImage,
-                        action: {})
-                    DrawerButton(
-                        title: Constants.SideDrawer.classesText,
-                        icon: Constants.SideDrawer.classesImage,
-                        action: {
-                        })
-                    DrawerButton(
-                        title: Constants.SideDrawer.faqText,
-                        icon: Constants.SideDrawer.faqImage,
-                        action: {
-                        })
-                    DrawerButton(
-                        title: Constants.SideDrawer.aboutText,
-                        icon: Constants.SideDrawer.aboutImage,
-                        action: {
-                        })
-                    DrawerButton(
-                        title: Constants.SideDrawer.signInText,
-                        icon: Constants.SideDrawer.signInImage,
-                        action: {
-                        })
-                    Spacer()
-                }
-                .frame(width: Constants.SideDrawer.frameWidth)
-                .background(
-                    RadialGradient(
-                        gradient: Gradient(colors: [
-                            Constants.SideDrawer.innerCircleColorRadius,
-                            Constants.SideDrawer.outerCircleColorRadius
-                        ]),
-                        center: .center,
-                        startRadius: 44,
-                        endRadius: 444
-                    )
-                )
-                .offset(
-                    x: viewModel.viewState.isMenuOpen ?
-                    0 : -(Constants.SideDrawer.frameWidth + viewModel.viewState.dragOffset.width)
-                )
-                .gesture(
-                    DragGesture(minimumDistance: 0)
-                        .onChanged { value in
-                            viewModel.updateDragOffset(value.translation)
-                        }
-                        .onEnded { _ in
-                            viewModel.endDragGesture()
-                        }
-                )
-                .animation(
-                    .easeInOut(duration: Constants.SideDrawer.animationDuration),
-                    value: viewModel.viewState.isMenuOpen)
-
-                Spacer()
-            }
+            SideDrawerView(viewModel: sideDrawerViewModel)
         }
     }
 }
