+++ BodyPositiveZumba/SideDrawer/Models/SideBarMenuItem.swift
@@ -1,0 +1,12 @@
+//
+//  SideBarMenuItem.swift
+//  BodyPositiveZumba
+//
+//  Created by Andrew Elbert on 1/27/25.
+//
+
+struct SideBarMenuItem {
+    let title: String
+    let icon: String
+    let action: () -> Void
+}
